# nest/src/projects/projects.controller.ts
import {
  Body,
  Controller,
  ForbiddenException,
  Get,
  Param,
  Patch,
  Post,
  UseGuards,
} from '@nestjs/common';
import { CurrentUser } from '../auth/current-user.decorator';
import { JwtAuthGuard } from '../auth/jwt-auth.guard';
import type { AuthenticatedUser } from '../auth/jwt.strategy';
import { ProjectAccessGuard } from '../auth/project-access.guard';
import { SlackService } from '../slack/slack.service';
import { CreateProjectDto } from './dto/create-project.dto';
import { UpdateProjectDto } from './dto/update-project.dto';
import { ProjectAccessService } from './project-access.service';
import { Project } from './project.entity';
import { ProjectsService } from './projects.service';

/** フロント返却用のプロジェクト形。Webhook URL の生値は出さず、設定有無のみ返す。 */
type ProjectResponse = Omit<Project, 'slackWebhookUrl' | 'tenant'> & {
  slackWebhookConfigured: boolean;
};

const toResponse = (p: Project): ProjectResponse => {
  // slackWebhookUrl は生値を出さない。tenant はリレーションなのでレスポンスから除く。
  const { slackWebhookUrl, tenant, ...rest } = p;
  void tenant;
  return { ...rest, slackWebhookConfigured: Boolean(slackWebhookUrl) };
};

const touchesSlack = (dto: UpdateProjectDto): boolean =>
  dto.slackWebhookUrl !== undefined ||
  dto.slackNotifyTaskCreated !== undefined ||
  dto.slackNotifyStatusChanged !== undefined ||
  dto.slackNotifyTaskCompleted !== undefined;

@Controller('projects')
@UseGuards(JwtAuthGuard, ProjectAccessGuard)
export class ProjectsController {
  constructor(
    private readonly projects: ProjectsService,
    private readonly access: ProjectAccessService,
    private readonly slack: SlackService,
  ) {}

  @Get()
  async list(@CurrentUser() user: AuthenticatedUser): Promise<ProjectResponse[]> {
    const rows = await this.projects.listByTenant(
      user.tenantId,
      await this.access.accessibleProjectIds(user),
    );
    return rows.map(toResponse);
  }

  @Post()
  async create(
    @CurrentUser() user: AuthenticatedUser,
    @Body() dto: CreateProjectDto,
  ): Promise<ProjectResponse> {
    const created = await this.projects.create(user.tenantId, dto);
    // 明示付与運用のため、作られたばかりのプロジェクトは作成者だけに見えるようにする
    await this.access.grant(user.tenantId, user.userId, created.id);
    return toResponse(created);
  }

  @Patch(':projectId')
  async update(
    @CurrentUser() user: AuthenticatedUser,
    @Param('projectId') id: string,
    @Body() dto: UpdateProjectDto,
  ): Promise<ProjectResponse> {
    // アーカイブ/復元 (archived の切替) は admin のみ許可
    if (dto.archived !== undefined && user.role !== 'admin') {
      throw new ForbiddenException('プロジェクトのアーカイブ/復元は管理者のみ実行できます');
    }
    // Slack 設定はプロジェクト管理者 or テナント管理者のみ
    if (touchesSlack(dto)) {
      await this.assertSlackAdmin(user, id);
    }
    return toResponse(await this.projects.update(user.tenantId, id, dto));
  }

  /** 設定画面の「テスト送信」。プロジェクト管理者 or テナント管理者のみ。 */
  @Post(':projectId/slack/test')
  async testSlack(
    @CurrentUser() user: AuthenticatedUser,
    @Param('projectId') id: string,
  ): Promise<{ ok: true }> {
    await this.assertSlackAdmin(user, id);
    return this.slack.sendTest(user.tenantId, id);
  }

  private async assertSlackAdmin(user: AuthenticatedUser, projectId: string): Promise<void> {
    if (user.role === 'admin') return;
    if (await this.projects.isProjectAdmin(projectId, user.userId)) return;
    throw new ForbiddenException('Slack 設定はプロジェクト管理者のみ変更できます');
  }
}