# nest/src/public/openapi/public-api.v1.yaml
openapi: 3.0.3
info:
  title: Backひろ 公開API
  version: 1.0.0
  description: |
    Backひろ の公開API。読み取りに加え、プロジェクトの登録・更新に対応する。

    ## 認証
    APIキー認証。`Authorization: Bearer <api key>` ヘッダを付与する。
    この公開API自体はトークンを発行しない。
    キーはアプリ画面（右上ユーザーメニュー →「APIキー」。admin / power_user のみ）で発行する。
    平文は発行時に一度だけ表示され、サーバには sha256 ハッシュのみ保存される。
    キーはユーザーに紐づき、アクセス範囲はそのユーザーの所属テナント・権限に固定される。

    ## プロジェクトの閲覧範囲
    キー所有ユーザーが閲覧を許可されたプロジェクトだけを操作できる（アプリ画面の
    ユーザー管理で admin が設定する）。許可されていないプロジェクトは
    一覧に出ず、`key` を直接指定しても `404` を返す（存在を伏せるため `403` にしない）。
    テナント admin のキーはこの設定に関係なく全プロジェクトを操作できる。
    `POST /v1/projects` で作成したプロジェクトは、作成者にのみ自動で閲覧権が付く。

    ## 識別子
    - プロジェクトは `key`（例: `KAISYU`）で識別する。
    - タスクはプロジェクト内連番 `seq`（表示用 `#N`）で識別する。
    - 内部 UUID は公開しない。ステータス等のマスタは `code` のまま返す。
      `code` / メンバー `id` → 表示名の解決は各マスタ取得エンドポイント
      （statuses / priorities / tags / flags / members / departments）を参照する。

    ## レート制限
    全体に IP 単位のレート制限あり（超過時 429）。
servers:
  - url: '{origin}/api'
    description: API ベース URL（グローバルプレフィックス /api）
    variables:
      origin:
        default: https://back-hiro.armabianca.com
        description: 'スキーム + ホスト（例 https://app.example.com）'
security:
  - ApiKeyAuth: []
...
paths:
  /v1/projects:
    get:
      tags: [Projects]
      summary: プロジェクト一覧
      description: APIキー所有ユーザーのテナントの、アーカイブされていないプロジェクトを返す。
      operationId: listProjects
      responses:
        '200':
          description: プロジェクト一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    post:
      tags: [Projects]
      summary: プロジェクト作成
      description: プロジェクトを作成する。key はテナント内一意（大文字化して保存）。重複時は 409。
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: 作成したプロジェクト
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/projects/{key}:
    patch:
      tags: [Projects]
      summary: プロジェクト更新（部分更新）
      description: 指定したフィールドのみ更新する。公開APIから更新できるのは name / description のみ。
      operationId: updateProject
      parameters:
        - $ref: '#/components/parameters/ProjectKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: 更新後のプロジェクト
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /v1/projects/{key}/tasks/{seq}:
    get:
      tags: [Tasks]
      summary: タスク取得
      description: プロジェクト内連番（seq）でタスク 1 件を取得する。
      operationId: getTask
      parameters:
        - $ref: '#/components/parameters/ProjectKey'
        - name: seq
          in: path
          required: true
          description: プロジェクト内連番（表示用 #N）。
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: タスク
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    patch:
      tags: [Tasks]
      summary: タスク更新（部分更新）
      description: |
        指定したフィールドのみ更新する。ステータス・担当者はここでは更新できない
        （POST /tasks/{seq}/status・POST /tasks/{seq}/member を使う）。
        links / tagCodes / flagCodes は送られた配列で全置換される（一部だけ送ると残りは消える）。
      operationId: updateTask
      parameters:
        - $ref: '#/components/parameters/ProjectKey'
        - $ref: '#/components/parameters/TaskSeq'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: 更新後のタスク
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /v1/projects/{key}/tags:
    get:
      tags: [Masters]
      summary: タグマスタ一覧
      description: タスクの tagCodes を解決するためのタグ一覧。
      operationId: listTags
      parameters:
        - $ref: '#/components/parameters/ProjectKey'
      responses:
        '200':
          description: タグ一覧
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Tag' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /v1/projects/{key}/flags:
    get:
      tags: [Masters]
      summary: フラグマスタ一覧
      description: タスクの flagCodes を解決するためのフラグ一覧。
      operationId: listFlags
      parameters:
        - $ref: '#/components/parameters/ProjectKey'
      responses:
        '200':
          description: フラグ一覧
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Flag' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /v1/projects/{key}/members:
    get:
      tags: [Masters]
      summary: メンバー一覧
      description: タスクの assigneeMemberId / requesterMemberId を解決するためのメンバー一覧。
      operationId: listMembers
      parameters:
        - $ref: '#/components/parameters/ProjectKey'
      responses:
        '200':
          description: メンバー一覧
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Member' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer <api key>`。キーは bh_live_ で始まる。'
  parameters:
    TaskSeq:
      name: seq
      in: path
      required: true
      description: 'プロジェクト内連番（表示用 #N）。'
      schema: { type: integer, minimum: 1 }
    ProjectKey:
      name: key
      in: path
      required: true
      description: プロジェクトの key（大文字小文字は区別しない／例 KAISYU）。
      schema: { type: string }
  responses:
    BadRequest:
      description: リクエスト不正（クエリのバリデーションエラー等）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: APIキー未指定または無効
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: プロジェクトまたはタスクが見つからない
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: 競合（プロジェクト key の重複など）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooManyRequests:
      description: レート制限超過
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Project:
      type: object
      required: [key, name, description, archived]
      properties:
        key:
          type: string
          example: KAISYU
        name:
          type: string
          example: システム改修対応
        description:
          type: string
          nullable: true
        archived:
          type: boolean
          description: アーカイブ済みか（一覧は非アーカイブのみ返す）
    ProjectCreate:
      type: object
      required: [key, name]
      properties:
        key:
          type: string
          maxLength: 64
          description: テナント内一意の識別子。大文字化して保存される。
          example: KAISYU
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 2000
          nullable: true
    ProjectUpdate:
      type: object
      description: 指定したフィールドのみ更新する（部分更新）。
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 2000
          nullable: true
...
    TaskUpdate:
      type: object
      description: |
        指定したフィールドのみ更新する（部分更新）。省略したフィールドは変更されない。
        null を指定できるフィールドは null で「未設定」に戻せる。
        links / tagCodes / flagCodes は送られた配列で全置換される（一部だけ送ると残りは消える）。
      properties:
        content:
          type: string
          maxLength: 500
          description: タスクのタイトル（一覧に表示される概要）。
        description:
          type: string
          maxLength: 50000
          description: 詳細説明。Markdown で記述できる。空文字で説明を消せる。
        links:
          type: array
          maxItems: 20
          items: { $ref: '#/components/schemas/TaskLink' }
          description: 関連リンク（ラベル + URL）。**全置換**なので現在のリンクも含めて全件送ること。空配列で全削除。最大 20 件。
        priorityCode:
          type: string
          nullable: true
          description: 優先度コード。`GET /projects/{key}/priorities` の code を指定する。null で未設定に戻す。
        requesterMemberId:
          type: string
          format: uuid
          nullable: true
          description: 起票者メンバーの id。`GET /projects/{key}/members` の id を指定する。プロジェクト外のメンバーは 400。null で未設定に戻す。
        requestingDeptCode:
          type: string
          nullable: true
          description: 依頼部署コード。`GET /departments` の code を指定する。null で未設定に戻す。
        deadline:
          type: string
          format: date
          nullable: true
          description: 期限（YYYY-MM-DD）。null で未設定に戻す。
        plannedStartDate:
          type: string
          format: date
          nullable: true
          description: 着手予定日（YYYY-MM-DD）。null で未設定に戻す。
        plannedCompletionDate:
          type: string
          format: date
          nullable: true
          description: 完了予定日（YYYY-MM-DD）。null で未設定に戻す。
        plannedReleaseDate:
          type: string
          format: date
          nullable: true
          description: リリース予定日（YYYY-MM-DD）。null で未設定に戻す。
        tagCodes:
          type: array
          items: { type: string }
          description: タグコードの配列。**全置換**なので現在のタグも含めて全件送ること。空配列で全解除。存在しないコードは黙って無視される。
        flagCodes:
          type: array
          items: { type: string }
          description: フラグコードの配列。**全置換**なので現在のフラグも含めて全件送ること。空配列で全解除。存在しないコードは黙って無視される。
...
    TaskLink:
      type: object
      required: [label, url]
      properties:
        label: { type: string }
        url: { type: string, format: uri }
    Task:
      type: object
      required:
        - seq
        - shortCode
        - content
        - description
        - links
        - statusCode
        - priorityCode
        - assigneeMemberId
        - requesterMemberId
        - requestingDeptCode
        - tagCodes
        - flagCodes
        - deadline
        - plannedStartDate
        - plannedCompletionDate
        - plannedReleaseDate
        - completedAt
        - statusChangedAt
        - createdAt
        - updatedAt
      properties:
        seq:
          type: integer
          description: プロジェクト内連番（表示用 #N）
          example: 15
        shortCode:
          type: string
          description: 共有リンク用の不透明な短縮コード（グローバル一意）
        content:
          type: string
          description: 一覧用の概要 / タイトル
        description:
          type: string
          description: 詳細説明（Markdown）
        links:
          type: array
          items: { $ref: '#/components/schemas/TaskLink' }
        statusCode:
          type: string
        priorityCode:
          type: string
          nullable: true
        assigneeMemberId:
          type: string
          format: uuid
          nullable: true
          description: 担当メンバーの ID（ProjectMember）。User ではない。
        requesterMemberId:
          type: string
          format: uuid
          nullable: true
        requestingDeptCode:
          type: string
          nullable: true
        tagCodes:
          type: array
          items: { type: string }
        flagCodes:
          type: array
          items: { type: string }
        deadline:
          type: string
          format: date
          nullable: true
        plannedStartDate:
          type: string
          format: date
          nullable: true
        plannedCompletionDate:
          type: string
          format: date
          nullable: true
        plannedReleaseDate:
          type: string
          format: date
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: 完了日時（ステータスが完了扱いの間だけ値を持つ）
        statusChangedAt:
          type: string
          format: date-time
          description: 最新ステータス変更日時
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Tag:
      type: object
      required: [code, name, color]
      properties:
        code: { type: string }
        name: { type: string }
        color: { type: string }
    Flag:
      type: object
      required: [code, name, color]
      properties:
        code: { type: string }
        name: { type: string }
        color: { type: string }
    Member:
      type: object
      required: [id, displayName]
      properties:
        id:
          type: string
          format: uuid
          description: メンバー ID（assigneeMemberId / requesterMemberId が参照）
        displayName: { type: string }
...
    Error:
      type: object
      properties:
        statusCode: { type: integer, example: 401 }
        message:
          oneOf:
            - type: string
            - type: array
              items: { type: string }
        error: { type: string, example: Unauthorized }